src/main/java/com/example/CrudAlunos/service/AlunoService/CadastrarAluno/CadastrarAlunoService.java
package com.example.CrudAlunos.service.AlunoService.CadastrarAluno;

import com.example.CrudAlunos.dto.AlunoDTO;
import com.example.CrudAlunos.model.Aluno;
import com.example.CrudAlunos.repository.AlunoRepository;
import com.example.CrudAlunos.repository.CursoRepository;

import java.util.Optional;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import java.util.logging.Logger;

@Service
public class CadastrarAlunoService {
    public static final Logger logger = Logger.getLogger(CadastrarAlunoService.class.getName());

    @Autowired
    private AlunoRepository alunoRepository;

    @Transactional
    public Aluno cadastrarAluno(AlunoDTO alunoDTO) {
        Aluno aluno = new Aluno();
        aluno.setNome(alunoDTO.getNome());
        aluno.setCpf(alunoDTO.getCpf());
        aluno.setMatricula(alunoDTO.getMatricula());
        return alunoRepository.save(aluno);
    }
}
